@@ -3,8 +3,9 @@ package com.marcoscg.licenser;
 import android.content.Context;
 import android.content.res.Resources;
 import android.graphics.Color;
-import android.support.annotation.ColorInt;
 import android.util.TypedValue;
+
+import androidx.annotation.ColorInt;
 
 /**
  * Created by @MarcosCGdev on 11/02/2018.
